Refactor All_stats function to include parallel computing option
The code changes in this commit refactor the All_stats function to include a new parameter, "parallel", which allows users to enable or disable parallel computing. When parallel computing is enabled, the code sets up a cluster and distributes the computations across multiple cores. This can significantly speed up the processing of large datasets with more than 1000 metabolites.

The changes also update the documentation and examples to reflect the new parameter and provide instructions on how to use it.

Note: The default value for "parallel" is set to FALSE, so parallel computing is disabled by default. Users need to explicitly set "parallel = TRUE" if they want to enable it.